+++ src/app/scss/_animations.scss
