+++ src/routes/Routes.tsx
@@ -1,21 +1,21 @@
 import * as React from "react";
 import { Route, Switch } from "react-router-dom";
 
-import { CheckoutLogin, NotFound } from "..";
+import { CheckoutLogin, NotFound } from "../components";
 import UserAccount, {
   baseUrl as userAccountBaseUrl
-} from "../../userAccount/routes";
-import { OrderDetails } from "../../userAccount/views";
-import { ArticlePage } from "../../views/Article";
-import { CartPage } from "../../views/Cart";
-import { CategoryPage } from "../../views/Category";
-import { CollectionPage } from "../../views/Collection";
-import { HomePage } from "../../views/Home";
-import OrderConfirmation from "../../views/OrderConfirmation/View";
-import { ProductPage } from "../../views/Product";
-import SearchPage from "../../views/Search";
+} from "../userAccount/routes";
+import { OrderDetails } from "../userAccount/views";
+import { ArticlePage } from "../views/Article";
+import { CartPage } from "../views/Cart";
+import { CategoryPage } from "../views/Category";
+import { CollectionPage } from "../views/Collection";
+import { HomePage } from "../views/Home";
+import OrderConfirmation from "../views/OrderConfirmation/View";
+import { ProductPage } from "../views/Product";
+import SearchPage from "../views/Search";
 
-const slugUrl = ':slug([a-z-0-9]+)/:id([0-9]+)/';
+const slugUrl = ":slug([a-z-0-9]+)/:id([0-9]+)/";
 export const baseUrl = "/";
 export const searchUrl = `${baseUrl}search/`;
 export const categoryUrl = `${baseUrl}category/${slugUrl}`;
@@ -43,3 +43,5 @@ export const Routes: React.FC = () => (
     <Route component={NotFound} />
   </Switch>
 );
+
+export default Routes;
