+++ src/routes/index.ts
@@ -1,0 +1,2 @@
+export * from "./Routes";
+export { default as Routes } from "./Routes";
